Fixed timestamp bug. Courtesy @lalith-AE30

+++ client.py
@@ -71,6 +71,6 @@ if __name__ == "__main__":
                     description='Headless client monitor')
     parser.add_argument('--host', type=str, default='localhost');           # positional argument
     parser.add_argument('-p', '--port', type=int, default=6444);
-    parser.add_argument('-r', '--rperiod', type=int, default=15);
+    parser.add_argument('-r', '--rperiod', type=float, default=15);
     args = parser.parse_args();
     Client(args.rperiod, args.host, args.port).begin();
